모임 x,y 자료형 변경

+++ src/main/java/kr/co/ouoe/MeetingPost/domain/MeetingLocate.java
@@ -18,10 +18,10 @@ public class MeetingLocate {
     //private Long meetingPostId;
 
     @Column(name = "lantitude",nullable = false)
-    private Long x;
+    private double x;
 
     @Column(name = "longitude",nullable = false)
-    private Long y;
+    private double y;
 
 
     @Builder
